==> TPS-2021/april/2. get_filenames_by_prefix.py <==
import tensorflow as tf
import os 
import pprint
import numpy as np

#Read csv file and convert it to tfrecord file
source_dir = "./customize_generate_csv/"
print(os.listdir(source_dir))

#Define file classification function
def get_filenames_by_prefix(source_dir, prefix_name):
    all_files = os.listdir(source_dir)
    results = []
    for filename in all_files:
        if filename.startswith(prefix_name):
            results.append(os.path.join(source_dir, filename))
    return results

train_filenames = get_filenames_by_prefix(source_dir, "train")
valid_filenames = get_filenames_by_prefix(source_dir, "valid")
test_filenames = get_filenames_by_prefix(source_dir , "test")

pprint.pprint(train_filenames)
pprint.pprint(valid_filenames)
pprint.pprint(test_filenames)
